refactor: Rename `lib/deployProfile.js`
