+++ lib/deploy-profile.test.js
@@ -19,7 +19,7 @@ describe('configureDeployProfile', () => {
       providerCredentials: { secretValue: { accessKeyId: 'id', secretAccessKey: 'secret' } },
     });
 
-    ({ configureDeployProfile } = proxyquire('./deployProfile', {
+    ({ configureDeployProfile } = proxyquire('./deploy-profile', {
       '@serverless/platform-client': { ServerlessSDK },
       './client-utils': {
         getPlatformClientWithAccessKey: () => ({
