+++ lib/plugin.js
@@ -16,7 +16,7 @@ const removeDestination = require('./removeDestination');
 const { saveDeployment, createAndSetDeploymentUid } = require('./deployment');
 const variables = require('./variables');
 const { generate } = require('./generateEvent');
-const { configureDeployProfile } = require('./deployProfile');
+const { configureDeployProfile } = require('./deploy-profile');
 const { test } = require('./test');
 const { getDashboardUrl, dashboardHandler } = require('./dashboard');
 const setApiGatewayAccessLogFormat = require('./setApiGatewayAccessLogFormat');
