+++ lib/plugin.test.js
@@ -49,7 +49,7 @@ describe('plugin', () => {
       './removeDestination': removeDestination,
       './deployment': { saveDeployment, createAndSetDeploymentUid },
       './generateEvent': { eventDict: {}, generate },
-      './deployProfile': { configureDeployProfile },
+      './deploy-profile': { configureDeployProfile },
       './injectLogsIamRole': injectLogsIamRole,
       './isAuthenticated': () => true,
       './setApiGatewayAccessLogFormat': setApiGatewayAccessLogFormat,
